Delete et quantités fonctionnel \o/

+++ src/Controller/BouquetCatController.php
@@ -18,7 +18,7 @@ class BouquetCatController extends AbstractController
     }
 
     /**
-     * Add into joint table
+     * Add or update quantity
      *
      * @param int $unit
      */
@@ -26,16 +26,30 @@ class BouquetCatController extends AbstractController
     {
         $idConcept = $_SESSION['id_bouquet_concept'];
         $bouquetCatManager = new BouquetCatManager();
-        $bouquetCatManager->insert($idConcept, $unit);
+
+        if (empty($bouquetCatManager->unitInConcept($unit))) {
+            $bouquetCatManager->insert($idConcept, $unit);
+        } else {
+            $bouquetCatManager->updateQuantUp($idConcept, $unit);
+        }
 
         header('location: /Concept/show/' . $_SESSION['id_bouquet_concept']);
     }
 
+    /**
+     * Delete or update quantity
+     *
+     * @param int $unit
+     */
     public function delete(int $unit)
     {
         $idConcept = $_SESSION['id_bouquet_concept'];
         $bouquetCatManager = new BouquetCatManager();
-        $bouquetCatManager->delete($idConcept, $unit);
+        if ($bouquetCatManager->unitInConcept($unit)[0] > 1) {
+            $bouquetCatManager->updateQuantDwn($idConcept, $unit);
+        } else {
+            $bouquetCatManager->delete($idConcept, $unit);
+        }
 
         header('location: /Concept/show/' . $_SESSION['id_bouquet_concept']);
     }
